refactor: Move code from main.go to other files
* Card ID: CCT-950
* Initial step of refactoring main.go.
* Code moved from main.go to other files
* Introduced new files for connect and disconnect
  commands
* Added several other files for interactive mode,
  logging
* Some code was moved to existing files
* Renamed register.go to rhsm.go, because it contains code
  not only for registration
