+++ canonical_facts_cmd.go
@@ -1,0 +1,23 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"github.com/urfave/cli/v2"
+)
+
+// canonicalFactAction tries to gather canonical facts about system,
+// and it prints JSON with facts to stdout.
+func canonicalFactAction(_ *cli.Context) error {
+	// NOTE: CLI context is not useful for anything
+	facts, err := GetCanonicalFacts()
+	if err != nil {
+		return cli.Exit(err, 1)
+	}
+	data, err := json.MarshalIndent(facts, "", "   ")
+	if err != nil {
+		return err
+	}
+	fmt.Println(string(data))
+	return nil
+}
