+++ logging.go
@@ -1,0 +1,8 @@
+package main
+
+import "github.com/subpop/go-log"
+
+type LogMessage struct {
+	level   log.Level
+	message error
+}
